META-270: Stub plugin: moved some settings from ini file to GUI
META-270: Stub settings improvements
META-283: AnalyticsDb: fixed blinking test AnalyticsDbCursor.interleaved_tracks_are_reported_interleaved
META-283: AnalyticsDb: fixed generation of off-screen object in some tests
META-283: Removed not used #include
META-285: Old plugin settings model stays up after reconnecting to a server without plugins
META-292: Add objectHashId to the binary database
META-292: Fix binary search for descending sort order
META-292: broken UT fixed
META-292: keep detailed timestamp for each matched objectGroupId
META-295: Analytics DB: Object lookup loop made finite explicitely
META-295: Analytics DB: dropped 'object_search' table
META-295: Analytics DB: filtering objects in memory after lookup
META-295: Analytics DB: filtering objects in-memory after selecting archive lookup
META-295: Analytics DB: fixed cursor tests
META-295: Analytics DB: fixed object lookup with a limit
META-295: Analytics DB: fixed selected object order when there is a limit
META-295: Analytics DB: fixed some bugs when looking up objects
META-295: Analytics DB: fixed some bugs when selecting objects from archive
META-295: Analytics DB: matching objects in the analytics archive first.
META-295: Compilation fix
META-295: fixed test for the archive with a single record
META-298: nx_kit: Introduce fromString() in nx::kit::utils.h
META-298: nx_kit: Removed unused function declaration
META-299: Fix MacOS installer background in the MetaVms customization
META-300: Fix displaying object frames in the zoom window
META-303: SDK: Provide Plugin Home Directory path to plugins
META-31: Added an error code to accompany a plugin loading status
META-31: English grammar fix in the apidoc comment
META-31: Replaced usage of deprecated QnMediaServerConnection with modern rest::ServerConnection
VMS-11543: Added plugin loading status, improved UI/UX
VMS-11543: Added plugin sorting by loaded status and name; greyed not loaded plugins
VMS-11543: Loading Server plugins: improving diagnostics and comments, related refactoring
VMS-11543: Replaced plugin information stub with actual server request
VMS-11543: Server plugins: Extended PluginInfo for bad manifests and supported interfaces
VMS-11543: Slightly improve code
VMS-11543: Small code fix (should not alter any logic)
VMS-13837: Add apidoc for /ec2/setResourceParams; improve apidoc for /ec2/getResourceParams
VMS-13858: add missed message to the 'api/cookieLogin' rest method
VMS-13963: mapping for Dahua DH-SD49225T-HN added
VMS-14090, VMS-14091: A new way to calculate Lilin SD card chunks start timestamp and duration
VMS-14091: Compilation fix
VMS-14091: Remote archive synchronization: added chunk id to synchronization logs
VMS-14290: added new Dahua encoder
VMS-14296: Disable ICE validation when building windows installers
VMS-14302: deb package: Get rid of removing '/opt/deb' directory due to its absense
VMS-14306: Crash when stopping Windows service 2 sec after start
VMS-14340: Fixed the usage of an uninitialized variable that is responsible for transport RTP
VMS-14349: Do not return thumbnails if requested time is in archive hole
VMS-14353: Wrong recordingTimeStampPeriods result for large detail level
VMS-14357: webSocket assert crash on sending ping message
WEB-1011: fix API documentation for ifconfig method

<==== Mercurial data ====>
Branch: vms_4.0_gui

+++ vms/libs/common/src/recording/time_period_list.h
@@ -62,7 +62,7 @@ public:
     static QnTimePeriodList filterTimePeriodsAsc(
         Iterator itr,
         Iterator endItr,
-        int detailLevel,
+        qint64 detailLevel,
         bool keepSmallChunks,
         int limit)
     {
@@ -102,7 +102,7 @@ public:
     static QnTimePeriodList filterTimePeriodsDesc(
         Iterator itr,
         Iterator endItr,
-        int detailLevel,
+        qint64 detailLevel,
         bool keepSmallChunks,
         int limit)
     {
@@ -141,7 +141,7 @@ public:
     static QnTimePeriodList filterTimePeriods(
         Iterator itr,
         Iterator endItr,
-        int detailLevel,
+        qint64 detailLevel,
         bool keepSmalChunks,
         int limit,
         Qt::SortOrder sortOrder)
